python_dj/requirements.txt
pyaudio==0.2.14
numpy==1.24.3
python-socketio[client]==5.11.0
websocket-client==1.6.4
